Create lexicon25.py
I'm not really sure if this lexicon will help me retain the information entirely. but its a great reference point for retaining some of the information. studying involves reading, video, and writing it out. read, write, and visualize the subject matter you're studying and you should do well with any exam your studying. but you have to do it for at least an hour a day. I may be accepting a job offer soon but in the meantime its best i stay busy with building my own data repository as well as focusing on my own micro-data center.